// src/models/event-model.tsx
  export interface EventModel  {
      title:       string;
      price:       string;
      image:       string;
      address:     string;
      date:        Date;
      description: string;
      speakers:    string;
      capacity:    string;
      spots:       string;
      id:          string;
}